Allow overriding of the fpgApplication class
A new private variable "fpgApplicationClass" and supporting type
"TfpgApplicationClass" have been added to fpg_main to tell
fpgApplication() what class to use to create the application object.
This defaults to TfpgApplication. A procedure "UseApplicationClass"
has been added to set the class and check that the fpgApplication
object has not already been created.

Several protected and public methods of TfpgApplication have been
made virtual to give devs the opportunity to override and extend
them.